Separated files from index
